fix: cast result to double in order to correctly compare equals values

+++ openscenario/openscenario_interpreter/src/syntax/distance_condition.cpp
@@ -295,7 +295,7 @@ auto DistanceCondition::evaluate() -> Object
 
   return asBoolean(triggering_entities.apply([&](auto && triggering_entity) {
     results.push_back(distance(triggering_entity));
-    return rule(results.back(), value);
+    return rule(static_cast<double>(results.back()), value);
   }));
 }
 }  // namespace syntax
